Made time_since_last_change() more robust
The time_since_last_change() function would barf if the temp file describing the last change time was not present - e.g. if the script had never been run before. This didn't have any functional impact, but it messed up some of the information printed to the log.

This is now fixed so that the current unix time is returned if there's no record of the last change time.