Handle significant (but legitimate) offset changes
My logs show that very occasionally the offset for a reflector's ICMP timestamps changes by a significant amount, and then stays that way. My guess is that this is due to a clock update/correction on the reflector. When this happened my ICMP timestamp correction code would unfairly reject the reflector as bad and replace it with a new one, because it would keep checking subsequent results against the original offset until the reflector struck out.

I have updated the code to update the reflector's offset if the attempt to correct the results fails, such that the next result will be checked using the offset based on these new results. If the change was legitimate the next result from this reflector will be ok. If the change was not legitimate and the ICMP timestamps (and therefore our calculated offset) are jumping around all over the place, correction will fail again next time.