Print reflector list when reloading reflector pool
Added code to print the current reflector list when the reflector pool is reloaded. Also added code to store the time at which each reflector is added to the list (using the previously unused values in the %reflector_ips hash), so we can get some limited insight into which reflectors are the most reliable.