Prevent deadlock when replacing reflector pool
Hit another deadlock when replacing the reflector pool. I'm not 100% sure of the cause, but the only possibility I can see right now is that my previous approach for minimising the scope of the %reflector_ips lock in check_latency() did not actually work, and that we were holding the %reflector_ips lock when we tried to stop the sender/receiver threads (to send a log message to the syslog). This would cause a deadlock if one of those threads was waiting on the %reflector_ips lock.