Handle IFB interfaces correctly
The script expects regular non-virtual interfaces to be specified in the configuration file. It will then handle the necessary complications to ensure that bandwidth changes are applied correctly using UCI and tc (i.e. using the correct download/upload ingress/egress terminology, and applying changes to IFB interfaces where appropriate).

Added code to "correct" IFB interfaces in the config to regular interfaces, and print a warning.

Some other unrelated small corrections are lumped in with this change.