Increase value of $max_recent_bandwidth_usages
After running fine for 90 days the script suddenly died because it couldn't find a record of the bandwidth usage for a particular time slot:

Fri Jul 21 15:47:36.167 2023 [6150-0015312751]: FATAL ERROR: Failed to get upload bandwidth at 15:47:13.287. Earliest of 45 (max 45) bandwidth samples' start time = 15:47:13.464

It seems that in some strange and rare circumstances the number I calculate for $max_recent_bandwidth_usages is not sufficient.

This is going to be very tricky to debug if it only happens once every three months, so this patch just artificially increases $max_recent_bandwidth_usages slightly so we should be covered. I was actually already bumping it slightly, but evidently  that arbitrary bump of 2 wasn't enough, so I've increased it to 10.

As mentioned in the code there should be zero performance impact - just a *tiny* increase in memory usage.