Add decrease-after-decrease delay
A delay between decreases is needed to ensure that a decrease in SQM bandwidth has enough time to take effect before we consider
decreasing it again. This is especially important (maybe *only* important?) in the download direction, because it takes time for remote senders to realise that they need to slow down the rate at which they are sending packets. The delay is controlled by the new "decrease_delay_after_decrease" property.