Prevent redundant warnings when WAN is down
A warning is printed when we cannot read WAN interface data from /proc/net/dev. This happens when the WAN interface doesn't exist, which is normal for PPPOE connections when the PPPOE tunnel isn't connected e.g. when the internet connection is down, or the router has only just booted up.

With this commit a warning will only be printed if the internet connection is up. This prevents the logs from being clogged up with pointless warnings when there's a WAN outage.