Improve startup WAN interface check
The current WAN interface check during startup can fail if the interface has only just come up (e.g. during booting), which causes the script to die unnecessarily.

This patch switches to a different approach for the check, examining the output from "ip link show" to determine whether or not the interface is valid.